Return HTTP 404:NOT FOUND for non-existing entries, approach 2.1
This commit uses the second approach described below.

The commit shows one way to implement the second approach that is
quite invasive since it implies that we add Spring annotations to
the used exception class. Thus, not recommended.

Spring allows for different approaches on how to control the return of
HTTP codes to the requesting client:

1. Upgrade the return type by wrapping the return type in a Spring
   ResponseEntity<T> type which in turn includes the object and an
   http code.
   Might be seen as a little bit ugly but usually works fine.

2. Let Spring handle the http code by making your object throw an
   exception that Spring can map to appropriate HTTP error code.
   This second approach could be implemented in different ways,
   some nice some ugly or invasive.