Return HTTP 404:NOT FOUND for non-existing entries, approach 2.2
This commit uses the second approach described below.

The commit shows a different, less invasive way to implement the
second approach compared to the previous commit. It does not
'infect' the Exception (POJO) class with any Spring annotations.

Note that this impementation still returns a http:404 html page
suitable for humans, *not* for machine-machine communication.

Spring allows for different approaches on how to control the return of
HTTP codes to the requesting client:

1. Upgrade the return type by wrapping the return type in a Spring
   ResponseEntity<T> type which in turn includes the object and an
   http code.
   Might be seen as a little bit ugly but usually works fine.

2. Let Spring handle the http code by making your object throw an
   exception that Spring can map to appropriate HTTP error code.
   This second approach could be implemented in different ways,
   some nice some ugly or invasive.