Return HTTP 404 for non-existing entries, prefered implementation
This commit also uses the second approach described below but
with an implemetation better suited for machine-machine
communication (not formatted to be read by humans).

The commit shows an implementation using a helper class to format
the returned data.

Spring allows for different approaches on how to control the return of
HTTP codes to the requesting client:

1. Upgrade the return type by wrapping the return type in a Spring
   ResponseEntity<T> type which in turn includes the object and an
   http code.
   Might be seen as a little bit ugly but usually works fine.

2. Let Spring handle the http code by making your object throw an
   exception that Spring can map to appropriate HTTP error code.
   This second approach could be implemented in different ways,
   some nice some ugly or invasive.