Use HATEOAS package to add additional data in returned datastructures
Here we use package /org.springframework.hateoas/ that provides
support to modify returned data structures in REST communication. In
this example we add a (HATEOAS) link for each returned entry when
aksed for all entries. In addition, when asked for a range of entries
we also return a link to next group of entries.

NOTE: Here we cheat by letting a real domain class extend the hateoas
class where we normally should use a separate /representation class/
for the REST communication and let that /representation class/ extend
hateoas class.

class CustomerRepresentation extends Customer, ResourceSupport {
...
}